--- A0/types.h
@@ -1,6 +1,0 @@
-#ifndef TYPES_H
-#define TYPES_H
-
-
-
-#endif
